backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-PyMongo==3.0.1
pymongo==4.11.1
dnspython==2.7.0
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.1.3
Jinja2==3.1.5
click==8.1.8
gunicorn==21.2.0